refactor(dashboard): extract Mastio PKI sub-router (F-B-201 PR-8/10)
Move the three /proxy/pki routes (overview page + export-ca download +
rotate-ca with CSRF + approval-hook gate) out of the ~3500-LOC
mcp_proxy/dashboard/router.py god-object into a dedicated pki_routes.py
module, mounted via include_router.

Same pattern as PR-2..PR-7. CSRF gate, approval-hook intercept on
rotate-ca and the CA private-key handling all preserved verbatim.

F-B-201 progress: 8/10 PR.
